+++ src/client/app.css
@@ -7,12 +7,6 @@ body {
   margin: auto;
 }
 
-.slider {
-  margin: auto;
-  margin-top: 10%;
-  width: 80%;
-}
-
 .container {
   min-height: 500px;
   width: 100%;
