+++ src/client/components/LabelBlock/LabelBlock.css
@@ -2,7 +2,7 @@
 
 .title-area {
     border: solid;
-    text-align: left;
-    padding: 16px 0 16px 24px;
+    text-align: center;
+    padding: 16px 0;
     /* word-wrap: break-word; */
 }
