+++ src/client/components/Sliders/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './Sliders';
